fix(image viewer): fix cropped image and keep the box visible
When an image appears to view, the image maintains its original ratio.
Before this, even an image succeeded to appear, it lost some part of it.

When drag and drop an image to the drop box, the image will be displayed
below the box. When this happens, it keeps the drop box visible. It means
that when a new image wants to be viewed, the box to drop the new image is
clear to see.

This commit also improve browse button, leave element <center>, using
ID element as querySelector argument and rename the only javascript file.

Resolves: #3, #4
See also: #2